Move Utils into new utils folder. Adapt import paths.
